Committing logical delete in UserRestController.java

+++ src/main/java/com/example/carpooling/exceptions/EntityNotFoundException.java
@@ -1,7 +1,7 @@
 package com.example.carpooling.exceptions;
 
 public class EntityNotFoundException extends RuntimeException {
-    public EntityNotFoundException(String type, int id) {
+    public EntityNotFoundException(String type, Long id) {
         this(type, "id", String.valueOf(id));
     }
